requirements.txt
GitPython==3.1.43
Jinja2==3.1.2
MarkupSafe==2.1.3
PyYAML==6.0.1
Pygments==2.17.2
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
eval_type_backport==0.2.0
executing==2.0.1
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.12.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==4.12.0
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.3.post0
lightning==2.3.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
matplotlib==3.8.4
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.4
monai==1.3.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
onnx-simplifier==0.4.36
onnx==1.16.0
onnxruntime==1.17.3
opencv-python-headless==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.2
prompt-toolkit==3.0.42
protobuf==5.26.1
psutil==5.9.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow-hotfix==0.6
pyarrow==16.1.0
pycocotools==2.0.7
pydantic==2.7.2
pydantic_core==2.18.3
pyparsing==3.1.2
pyreadline3==3.4.1
python-box==7.2.0
python-dateutil==2.9.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.1
pytz==2024.1
pywin32==227
pyzmq==25.1.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
segment-anything==1.0
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.7
tomli==2.0.1
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchmetrics==1.4.0.post0
torchvision==0.17.2+cu118
tornado==6.2
tqdm==4.66.4
traitlets==5.14.2
typing_extensions==4.11.0
typing_extensions==4.8.0
tzdata==2024.1
ultralytics-thop==2.0.0
ultralytics==8.2.63
urllib3==2.2.1
wandb==0.17.4
wcwidth==0.2.13
wheel==0.41.2
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zipp==3.18.1